+++ CHANGES.md
@@ -1,5 +1,11 @@
 ## Changes
 
+2.5.1
+* Fixed bug where multiple headers where served as one response.
+* Fixed missing Public-Key-Pinning policy when not using .htaccess file.
+* Added support for Feature-Policy directive.
+* Tested with WordPress 5.0.2
+
 2.5.0
 * Complete rewrite of Carl Conrad's plugin from scratch.
 * Separated code from design where applicable.